tests: did some refactoring in test setup code to encapsulate stubbing of core internal yeoman-generator functions
